Started TEXT1

+++ SDHRWindow.h
@@ -5,7 +5,7 @@
 #include "shader.h"
 
 typedef struct uixy { uint32_t x; uint32_t y; } uXY;
-typedef struct ixy { int64_t x; int64_t y; } iXY;
+typedef struct ixy { int32_t x; int32_t y; } iXY;
 
 class SDHRWindow
 {
